+++ packages/projects/src/actions/projectTaskActions.ts
@@ -38,6 +38,7 @@ import {
 } from '../schemas/project.schemas';
 import { OrderingService } from '../lib/orderingUtils';
 import { buildProjectTaskWebhookChanges } from '../lib/projectTaskWebhookChanges';
+import { applyTicketLinkRestriction } from '../lib/taskTicketMapping';
 import { validateAndFixOrderKeys } from './regenerateOrderKeys';
 import {
   buildProjectTaskAssignedPayload,
@@ -422,26 +423,6 @@ export async function buildTaskAssigneeSetByTaskId(
     }
 
     return result;
-}
-
-/**
- * Returns a copy of a ticket-link row with non-essential fields nulled when
- * the caller isn't allowed to see the ticket in full. Caller is expected to
- * compute `allowedTicketIds` via filterAuthorizedTicketIds.
- */
-export function applyTicketLinkRestriction(
-    link: IProjectTicketLinkWithDetails,
-    allowedTicketIds: Set<string>
-): IProjectTicketLinkWithDetails {
-    if (allowedTicketIds.has(link.ticket_id)) {
-        return { ...link, restricted: false };
-    }
-    return {
-        ...link,
-        status_name: null,
-        is_closed: null,
-        restricted: true,
-    };
 }
 
 export async function filterAuthorizedTicketIds(
